requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
PyExecJS==1.5.1
PyMySQL==0.9.3
requests==2.22.0
six==1.12.0
urllib3==1.25.2
